+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # changelog
 
+## [0.2.0] - 2024-11-20
+### new features
+- add `dirty-change` event to notify user about dirty status change in form elements.   
+- add `validity-change` event to notify user about validation change in form elements.   
+
 ## [0.1.0] - 2024-10-10
 ### new features
 - add virtual element so non web-component element can register themselves into form and get validate and etc